college-admission-qa/requirements.txt
flask
nltk
scikit-learn
numpy
python-dotenv
